% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_cv.R
\name{build_cv}
\alias{build_cv}
\title{Build a CV using latex}
\usage{
build_cv(content, style, out = NULL, pub_score = TRUE, clean = TRUE)
}
\arguments{
\item{content}{Path to a YAML file which contains all your personal
information as well as the structure of your CV.}

\item{style}{Path to the latex style you wish to apply.}

\item{out}{The directory where your output should be stored (without trailing
forward or back slash).}

\item{pub_score}{If TRUE, the publications section will start with an
overview of the current publication record including number of
publications, h- and h10-factor. For more information, see
[get_pub_record].}

\item{clean}{If \code{TRUE}, all files but the output pdf will be deleted
from the output directory.}
}
\description{
Build a CV using latex
}
\details{
\code{build_cv} builds a CV using a predefined style and YAML files
  which contain positions held, employments, projects worked on, etc.
}
