% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCV.R
\name{CleanNames}
\alias{CleanNames}
\title{Apostophes cause problems. This will fix them.}
\usage{
CleanNames(citations)
}
\arguments{
\item{citations}{Bibtext formatted, presumably orcid.info$journals}
}
\value{
Same object, but with curly quotes and similar evil symbols fixed
}
\description{
Apostophes cause problems. This will fix them.
}

