% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost-functions.R
\name{mse}
\alias{mse}
\alias{BayesRule}
\alias{BayesRule2}
\title{Cost Functions for Fitted Regression Models}
\usage{
mse(y, yhat)

BayesRule(y, yhat)

BayesRule2(y, yhat)
}
\arguments{
\item{y}{response}

\item{yhat}{fitted value}
}
\description{
Compute cost functions (cross-validation criteria) for fitted
regression models.
}
\section{Functions}{
\itemize{
\item \code{mse()}: Mean-Square Error

\item \code{BayesRule()}: Bayes for a binary response

\item \code{BayesRule2()}: Bayes for a binary response (without bounds checking)

}}
