% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCV.R
\name{CreateEducationMarkdown}
\alias{CreateEducationMarkdown}
\title{Create a Markdown document of education from biographical info}
\usage{
CreateEducationMarkdown(orcid.info, outdir = tempdir())
}
\arguments{
\item{orcid.info}{The list of info from orcid}

\item{outdir}{The directory to store the markdown file in}
}
\description{
Create a Markdown document of education from biographical info
}

