% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCV.R
\name{CreateFundingMarkdown}
\alias{CreateFundingMarkdown}
\title{Create a Markdown document of funding from biographical info}
\usage{
CreateFundingMarkdown(orcid.info, outdir = tempdir(),
  additional.te = "This is all in addition to other **funding my students have gotten** (NSF EAPSI grant, fellowships from NIMBioS and PEER (an NIH-funded program at UTK), Google Summer of Code funding), **funding for workshops or working groups** (from NIMBioS and the Society for Systematic Biologists), and **funding I got before my faculty position** (NESCent postdoctoral fellowship, NSF DDIG, NSF GRF, and various internal grants at UC Davis).")
}
\arguments{
\item{orcid.info}{The list of info from orcid}

\item{outdir}{The directory to store the markdown file in}
}
\description{
Create a Markdown document of funding from biographical info
}

