% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCV.R
\name{CreateHeadMarkdown}
\alias{CreateHeadMarkdown}
\title{Copy the head markdown}
\usage{
CreateHeadMarkdown(outdir = tempdir())
}
\arguments{
\item{outdir}{The directory to store the markdown file in}
}
\description{
Copy the head markdown
}
