% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCV.R
\name{CreateMarkdown}
\alias{CreateMarkdown}
\title{Create a Markdown document from biographical info}
\usage{
CreateMarkdown(orcid.info = GetInfoFromOrcid(), outdir = ".")
}
\arguments{
\item{orcid.info}{The list of info from orcid}

\item{outdir}{The directory to store the markdown file in}
}
\description{
Create a Markdown document from biographical info
}

