% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCV.R
\name{CreatePeopleMarkdown}
\alias{CreatePeopleMarkdown}
\title{Create a Markdown document of people in the lab from biographical info}
\usage{
CreatePeopleMarkdown(infile = system.file("extdata", "people.txt", package =
  "cv"), outdir = tempdir())
}
\arguments{
\item{infile}{The path to the text delimited file}

\item{outdir}{The directory to store the markdown file in}
}
\description{
Create a Markdown document of people in the lab from biographical info
}
