% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCV.R
\name{CreatePresentationsMarkdown}
\alias{CreatePresentationsMarkdown}
\title{Copy the presentations markdown}
\usage{
CreatePresentationsMarkdown(outdir = tempdir())
}
\arguments{
\item{outdir}{The directory to store the markdown file in}
}
\description{
Copy the presentations markdown
}
