% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCV.R
\name{CreateServiceMarkdown}
\alias{CreateServiceMarkdown}
\title{Create a Markdown document of service from biographical info}
\usage{
CreateServiceMarkdown(infile = system.file("extdata", "service.txt",
  package = "cv"), outdir = tempdir())
}
\arguments{
\item{infile}{The path to the text delimited file}

\item{outdir}{The directory to store the markdown file in}
}
\description{
Create a Markdown document of service from biographical info
}
