% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCV.R
\name{CreateSummaryMarkdown}
\alias{CreateSummaryMarkdown}
\title{Create a Markdown document to summarize me}
\usage{
CreateSummaryMarkdown(orcid.info, outdir = tempdir(),
  publications.offset = -2,
  prominent.pubs = "*Science, Nature, Ann. Rev Ecology, Evolution & Systematics, Systematic Biology, Evolution*, etc.")
}
\arguments{
\item{orcid.info}{The list of info from orcid}

\item{outdir}{The directory to store the markdown file in}

\item{publications.offset}{How to change the publication count.

You may want an offset if some of your publications shouldn't count.
For example, I have two Nature "papers" that are actually corrigendia,
that is, corrections for errors. Shouldn't really count, I think.}
}
\description{
Create a Markdown document to summarize me
}
