% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCV.R
\name{CreateTeachingMarkdown}
\alias{CreateTeachingMarkdown}
\title{Copy the teaching markdown}
\usage{
CreateTeachingMarkdown(outdir = tempdir())
}
\arguments{
\item{outdir}{The directory to store the markdown file in}
}
\description{
Copy the teaching markdown
}
