% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv-package.R
\docType{package}
\name{cv-package}
\alias{cv}
\alias{cv-package}
\title{cv: Files and Data for my CV}
\description{
To store my CV data and to generate it.
}
\author{
\strong{Maintainer}: Luke Johnston \email{lwjohnst@gmail.com} (\href{https://orcid.org/0000-0003-4169-2616}{ORCID})

}
\keyword{internal}
