% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv-timeseries.R
\name{cv.gls}
\alias{cv.gls}
\alias{Arima}
\alias{print.ARIMA}
\alias{update.ARIMA}
\alias{coef.ARIMA}
\alias{predict.ARIMA}
\alias{cv.ARIMA}
\title{Cross-Validate Mixed-Effects Model}
\usage{
\method{cv}{gls}(
  model,
  data = insight::get_data(model),
  criterion = mse,
  k = 10L,
  reps,
  seed,
  criterion.name = deparse(substitute(criterion)),
  i.only = FALSE,
  details = k <= 10L,
  confint = n >= 400L,
  level = 0.95,
  ncores = 1L,
  ...
)

Arima(
  formula,
  data,
  subset = NULL,
  na.action = na.pass,
  order = c(1L, 0L, 0L),
  ...
)

\method{print}{ARIMA}(x, ...)

\method{update}{ARIMA}(object, ...)

\method{coef}{ARIMA}(object, ...)

\method{predict}{ARIMA}(object, n.ahead = 1L, newdata = NULL, se.fit = FALSE, ...)

\method{cv}{ARIMA}(
  model,
  data = model$data,
  criterion = mse,
  k = 10L,
  reps,
  seed,
  i.only = FALSE,
  criterion.name = deparse(substitute(criterion)),
  details = k <= 10L,
  ...
)
}
\arguments{
\item{model}{an object of class \code{"gls"} produced by the
\code{\link[nlme]{gls}()} function, or of class \code{"ARIMA"}
produced by the \code{Arima()} function.}

\item{data}{data frame with the data to which the model was fit;
can usually be inferred from the model; for \code{Arima()}, a data
frame with data to which the model is to be fit.}

\item{criterion}{function to compute the CV cost criterion
(default \code{\link{mse}}).}

\item{k}{number of folds, an integer or "n" or "loo" for
leave-one-out CV; the default is \code{10}.}

\item{reps}{ignored (to match \code{\link{cv}()} generic function.)}

\item{seed}{ignored (to match \code{\link{cv}()} generic function.)}

\item{criterion.name}{name of the CV criterion; can usually be
inferred from \code{criterion}.}

\item{i.only}{if \code{TRUE} (the default is \code{FALSE}), predict
the response for cases in the i-th fold from the model fit to data
in the preceding fold only rather than fit to data from \emph{all}
preceding folds.}

\item{details}{return fold-wise statistics, including the CV criterion
and parameter estimates for cases in each fold after the first;
the default is \code{TRUE} for \eqn{k \le 10}.}

\item{confint}{if \code{TRUE} (the default if \eqn{n \ge 400}), report
a confidence interval for the adjusted CV criterion.}

\item{level}{level for the confidence interval (default \code{0.95}).}

\item{ncores}{if \code{ncores} \eqn{> 1}, the computation is parallelized.}

\item{...}{further arguments to be passed to \code{\link{arima}()}
or \code{Arima()}.}

\item{formula}{either a one-sided formula giving the response variable
in an ARIMA model with no predictors, or a two-sided formula with the
response on the left and terms for the predictors on the right (as in
a typical R regression model).}

\item{subset}{subsetting expression.}

\item{na.action}{function to process missing data; the default,
\code{na.pass}, will pass missing data to the \code{\link{arima}()}
function.}

\item{order}{the \eqn{p, d, q} specification of the ARIMA model;
see \code{\link{arima}()} for details.}

\item{x}{an object of class \code{"ARIMA"}.}

\item{object}{an object of class \code{"ARIMA"}.}

\item{n.ahead}{number of future cases to predict; the default is \code{1}.}

\item{newdata}{data frame with \code{n.ahead} rows containing the
predictors (if any) for the predicted future cases.}

\item{se.fit}{if \code{TRUE} (the default is \code{FALSE}), compute
the standard errors of the predictions.}
}
\description{
\code{\link{cv}()} methods for time-series models of class \code{"gls"}, fit
by the \code{\link[nlme]{gls}()} in the \pkg{lme4} package, and
by \code{Arima()}, which provides a formula interface to the
\code{link{arima}()}) function.
}
\section{Functions}{
\itemize{
\item \code{cv(gls)}: \code{cv()} method for \code{\link[nlme]{gls}()} models from the \pkg{nlme} package.

\item \code{Arima()}: model-formula wrapper for the \code{\link{arima}()} function.

\item \code{print(ARIMA)}: \code{print()} method for \code{"ARIMA"} objects
created by the \code{\link{Arima}()} function.

\item \code{update(ARIMA)}: \code{update()} method for \code{"ARIMA"} objects
created by the \code{\link{Arima}()} function.

\item \code{coef(ARIMA)}: \code{coef()} method for \code{"ARIMA"} objects
created by the \code{\link{Arima}()} function.

\item \code{predict(ARIMA)}: \code{predict()} method for \code{"ARIMA"} objects
created by the \code{\link{Arima}()} function.

\item \code{cv(ARIMA)}: \code{cv()} method for \code{"ARIMA"} objects
created by the \code{\link{Arima}()} function.

}}
