% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_function.R
\name{cv_function}
\alias{cv_function}
\title{Calculate the coefficient of variation}
\usage{
cv_function(x, na.rm = TRUE)
}
\arguments{
\item{.x.}{The variable of interest we use from the dataset to calculate the coefficient of variation. I am using the label "x" to represent the variable of interest. I also chose to use "x" as this is common notation used in field of statistics.}

\item{.na.rm.}{This is a logical argument used to ensure that any missing values (NA) are removed from the dataset. I have used the label na.arm as this is the common label used in this field when defining what to do when there is missing data (i.e., NAs) in the dataset.}
}
\value{
A numeric value is returned that tells us what the coefficient of variation is.
}
\description{
This calculates the coefficient of variation (standard deviation divded by the mean) to assess variability in the dataset.
}
\examples{
cv(gapminder$lifeExp)
cv(gapminder$pop)
cv(gapminder$gdpPercap)

}
