% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCV.R
\name{FinalCompileCV}
\alias{FinalCompileCV}
\title{Compile a set of markdown documents and convert with pandoc}
\usage{
FinalCompileCV(input = c("head.md", "summary.md", "education.md",
  "employment.md", "publications.md", "teaching.md", "funding.md",
  "presentations.md", "people.md", "service.md"), outdir = tempdir(),
  css = system.file("extdata", "format.css", package = "cv"),
  output = "OMearaCV", open.files = TRUE)
}
\arguments{
\item{input}{Vector of markdown documents}

\item{outdir}{The directory to store the output in}

\item{css}{The css file with formatting info.}

\item{output}{The output base file name. You'll receive <output>.pdf and <output>.html files.}

\item{open.files}{If TRUE, open the output files}
}
\description{
Compile a set of markdown documents and convert with pandoc
}
