% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_cv.R
\name{format_address}
\alias{format_address}
\title{Formats the address for printing}
\usage{
format_address(address)
}
\arguments{
\item{address}{a list of address lines.}
}
\value{
a character including lines breaks for TeX
}
\description{
Formats the address for printing
}
