% Generated by roxygen2 (4.0.2): do not edit by hand
\name{generate_cv}
\alias{generate_cv}
\title{Generates a CV from given content and style files}
\usage{
generate_cv(content, style, outdir = "output")
}
\arguments{
\item{content}{a character giving the name of the YAML file
containing the vita content}

\item{style}{a character giving the name of the style file.}
}
\description{
Generates a CV from given content and style files
}

