% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pub_record.R
\name{get_pub_record}
\alias{get_pub_record}
\title{Get Jannes publication record}
\usage{
get_pub_record()
}
\value{
A list containing Jannes publication record
}
\description{
Get Jannes publication record
}
\details{
Scrape publication record from FSU homepage, WOS and Google Scholar
}
