% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCV.R
\name{GetCollaborators}
\alias{GetCollaborators}
\title{Get collaborators from past time period}
\usage{
GetCollaborators(orcid.info, starting.year = 2013, outdir = tempdir())
}
\arguments{
\item{orcid.info}{The list of info from orcid}

\item{starting.year}{What year to start including collaborators}

\item{outdir}{The directory to store the temporary markdown file in}
}
\value{
vector of the authors
}
\description{
Get collaborators from past time period
}
