% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{getResponse}
\alias{getResponse}
\alias{getResponse.default}
\title{Extract response variable(s)}
\usage{
getResponse(model, ...)

\method{getResponse}{default}(model, ...)
}
\arguments{
\item{model}{a fitted model}

\item{...}{additional parameters for specific methods}
}
\value{
a vector or matrix containing the values of response variable(s)
}
\description{
Generic function to extract the response variable(s) from a fitted model
}
\section{Methods (by class)}{
\itemize{
\item \code{getResponse(default)}: default method

}}
\examples{
    fit <- lm(cbind(hp, mpg) ~ gear, mtcars)
    getResponse(fit)
}
