% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.modList.R
\name{models}
\alias{models}
\alias{cv.modList}
\alias{print.cvModList}
\title{Cross-Validate Several Models Fit to the Same Data}
\usage{
models(...)

\method{cv}{modList}(model, data, criterion = mse, k, reps, seed, ...)

\method{print}{cvModList}(x, ...)
}
\arguments{
\item{...}{for \code{cv()}, additional arguments to be passed to the \code{cv()} method
applied to each model. For \code{models()}, two or more competing models fit to the
the same data; the several models may be named. For the \code{print()}
method, arguments to be passed to the \code{print()} method for
the individual model cross-validations.}

\item{data}{(required) the data set to which the models were fit.}

\item{k}{the number of CV folds; may be omitted, in which case the value
will depend on the default for the \code{cv()} method invoked for the
individual models.}

\item{reps}{ignored (present only to match the generic function).}

\item{seed}{(optional) seed for R's pseudo-random-number generator,
to be used to create the same set of CV folds for all of the models;
if omitted, a seed will be randomly generated and saved.}

\item{x}{an object of class \code{"cvModList"} to be printed.}

\item{name}{criterion the CV criterion (cost) function, defaults to
\code{\link{mse}}.}
}
\value{
\code{models()} returns a \code{"modList"} object, the
\code{cv()} method for which rturns a \code{"cvModList"} object.
}
\description{
A \code{\link{cv}()} method for an object of class  \code{"modlist"},
created by the \code{models()} function. This \code{cv()} method simplifies
the process of cross-validating several models on the same set of folds.
}
\details{
\code{models()} performs some
"sanity" checks, warning if the models are of differrent classes, and
reporting an error if they are fit to apparently different data sets or
different response variables.
, , =mse, k=10, reps, seed, ...
}
\section{Functions}{
\itemize{
\item \code{models()}: create a list of models

\item \code{cv(modList)}: \code{cv} method for \code{"modList} objects

\item \code{print(cvModList)}: \code{print()} method for \code{"cvModList} object

}}
\examples{
data("Duncan", package="carData")
m1 <- lm(prestige ~ income + education, data=Duncan)
m2 <- lm(prestige ~ income + education + type, data=Duncan)
m3 <- lm(prestige ~ (income + education)*type, data=Duncan)
cv(models(m1=m1, m2=m2, m3=m3), data=Duncan, seed=7949)
}
