% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{mse}
\alias{mse}
\alias{BayesRule2}
\alias{BayesRule}
\title{Cost functions for fitted models}
\usage{
mse(y, yhat)

BayesRule2(y, yhat)

BayesRule(y, yhat)
}
\arguments{
\item{y}{response}

\item{yhat}{fitted value}
}
\description{
Cost functions for fitted models
}
\section{Functions}{
\itemize{
\item \code{BayesRule2()}: Bayes Rule 2

\item \code{BayesRule()}: Bayes Rule 2

}}
