% Generated by roxygen2 (4.0.2): do not edit by hand
\name{process_yaml}
\alias{process_yaml}
\title{Processes a specified YAML file to create outputs}
\usage{
process_yaml(section, input_dir = "../content", output_dir = "output",
  format = "tex")
}
\arguments{
\item{section}{a character giving the name of the section to
process.  There must be a corresponding \code{section.yaml} file in
the input directory and \code{format_section} function in the
formats file.  The function should take a YAML parsed list as input
and returned a list of lines in the correct format.}

\item{input_dir}{the input directory (no trailing /)}

\item{output_dir}{the output directory (no trailing /)}

\item{format}{a character giving a file name of the format
\code{format_functions.r}.  The default is "tex"}
}
\description{
Processes a specified YAML file to create outputs
}

