% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-cv-items.R
\name{read_bib_items}
\alias{read_bib_items}
\title{Read in the \code{.bib} file of publications and research output.}
\usage{
read_bib_items()
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} of the research outputs and publications.
}
\description{
Read in the \code{.bib} file of publications and research output.
}
