% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-cv-items.R
\name{read_cv_items}
\alias{read_cv_items}
\title{Read in the CV dataset.}
\usage{
read_cv_items()
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} of all CV items.
}
\description{
Read in the CV dataset.
}
