% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tidy_dates}
\alias{tidy_dates}
\title{Tidy up date variables.}
\usage{
tidy_dates(.tbl)
}
\arguments{
\item{.tbl}{Dataset to tidy with \code{start} and \code{end} date variables.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} dataset with tidied date variables.
}
\description{
Tidy up date variables.
}
