% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCV.R
\name{CreatePublicationsMarkdown}
\alias{CreatePublicationsMarkdown}
\title{Create a Markdown document of funding from biographical info}
\usage{
CreatePublicationsMarkdown(orcid.info, outdir = tempdir(),
  emphasis.name = "O'Meara")
}
\arguments{
\item{orcid.info}{The list of info from orcid}

\item{outdir}{The directory to store the markdown file in}

\item{emphasis.name}{The name to bold in the publications list. Presumably your own.}
}
\description{
Create a Markdown document of funding from biographical info
}

