% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv-mixed-models.R
\name{cv.merMod}
\alias{cv.merMod}
\title{Cross-Validate Mixed-Effects Model}
\usage{
\method{cv}{merMod}(
  model,
  data = insight::get_data(model),
  criterion = mse,
  k,
  seed,
  ncores = 1,
  clusterVariables,
  includeRandom = TRUE,
  ...
)
}
\arguments{
\item{model}{a regression model object of class \code{"merMod"}.}

\item{data}{data frame to which the model was fit (not usually necessary)}

\item{criterion}{cross-validation criterion function of form \code{f(y, yhat)}
where \code{y} is the observed values of the response and
\code{yhat} the predicted values; the default is \code{\link{mse}}
(the mean-squared error)}

\item{k}{perform k-fold cross-validation; \code{k}
may be a number or \code{"loo"} or \code{"n"} for n-fold (leave-one-out)
cross-validation; the default is \code{10} if cross-validating individual
cases and \code{"loo"} if cross-validating clusters.}

\item{seed}{for R's random number generator; optional, if not
supplied a random seed will be selected and saved; not needed
for n-fold cross-validation}

\item{ncores}{number of cores to use for parallel computations
(default is \code{1}, i.e., computations aren't done in parallel)}

\item{clusterVariables}{a character vector of names of the variables
defining clusters for a mixed model with nested random effects;
if missing, cross-validation is performed for individual cases rather than
for clusters}

\item{includeRandom}{include the random effects in predicting cases
in the clusters used to fit the model with each fold deleted (default
is \code{TRUE})? Predictions for cases in the omitted clusters for
each fold are always based only on the fixed effects.}

\item{...}{to match generic}
}
\description{
A \code{\link{cv}()} method for models of class \code{"merMod"}, fit
by the \code{\link[lme4]{lmer}()} and \code{\link[lme4]{glmer}()} functions
in the \pkg{lme4} package. The implementation here should be regarded
as experimental.
}
\details{
For mixed-effects models fit by the \code{lmer()} or \code{glmer{}} functions
in the \strong{nlme} package, cross-validation can be done by "clusters" or by
individual observations. If the former, predictions are based only on fixed
effects; if the latter, predictions include the random effects. Only mixed
models with fully nested random effects are supported.
}
\examples{
library(lme4)
# from ?lmer:
(fm1 <- lmer(Reaction ~ Days + (Days | Subject), sleepstudy))
cv(fm1, clusterVariables="Subject") # LOO CV of clusters
cv(fm1, seed=447) # 10-fold CV of cases

}
\seealso{
\code{\link{cv}}, \code{\link[lme4]{lmer}}, \code{\link[lme4]{glmer}}
}
