% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv-mixed-models.R, R/cv.R
\name{cv.merMod}
\alias{cv.merMod}
\alias{cv}
\alias{cv.default}
\alias{print.cv}
\alias{cv.lm}
\alias{cv.glm}
\title{Cross-validate fitted model}
\usage{
\method{cv}{merMod}(
  model,
  data = insight::get_data(model),
  criterion = mse,
  k = nrow(clusters),
  seed,
  parallel = FALSE,
  ncores = parallelly::availableCores(logical = FALSE),
  clusterVariables,
  ...
)

cv(model, data, criterion, k, seed, ...)

\method{cv}{default}(
  model,
  data = insight::get_data(model),
  criterion = mse,
  k = nrow(data),
  seed,
  parallel = FALSE,
  ncores = parallelly::availableCores(logical = FALSE),
  ...
)

\method{print}{cv}(x, ...)

\method{cv}{lm}(
  model,
  data = insight::get_data(model),
  criterion = mse,
  k = nrow(data),
  seed,
  parallel = FALSE,
  ncores = parallelly::availableCores(logical = FALSE),
  ...
)

\method{cv}{glm}(
  model,
  data = insight::get_data(model),
  criterion = mse,
  k = nrow(data),
  seed,
  parallel = FALSE,
  ncores = parallelly::availableCores(logical = FALSE),
  approximate = FALSE,
  ...
)
}
\arguments{
\item{model}{a model object that responds to model.frame(), update(), and predict()
and for which the response is stored in model$y or accessible via model.response()}

\item{data}{data frame to which the model was fit (not usually necessary)}

\item{criterion}{cross-validation criterion function of form f(y.obs, y.fitted)
(default is mse)}

\item{k}{perform k-fold cross-validation (default is n-fold)}

\item{seed}{for R's random number generator}

\item{parallel}{do computations in parallel? (default is FALSE)}

\item{ncores}{number of cores to use for parallel computations
(default is number of physical cores detected)}

\item{clusterVariables}{a character vector of names of the variables
defining clusters for a mixed model with nested random effects}

\item{...}{to match generic}

\item{x}{a `cv` object to be printed}

\item{approximate}{perform approximate leave-k-out computation for a GLM}
}
\value{
a "cv" object with the cv criterion averaged across the folds,
the bias-adjusted averaged cv criterion,
the criterion applied to the model fit to the full data set,
and the initial value of R's RNG seed
}
\description{
Generic function to cross-validate fitted models using fast algorithms
}
\section{Methods (by class)}{
\itemize{
\item \code{cv(merMod)}: merMod method

\item \code{cv(default)}: default method

\item \code{cv(lm)}: lm method

\item \code{cv(glm)}: glm method

}}
\section{Methods (by generic)}{
\itemize{
\item \code{print(cv)}: print method

}}
