% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv-mixed-models.R, R/cv.R
\name{cv.merMod}
\alias{cv.merMod}
\alias{cv}
\alias{cv.default}
\alias{print.cv}
\alias{cv.lm}
\alias{cv.glm}
\title{Cross-Validate Regression Models}
\usage{
\method{cv}{merMod}(
  model,
  data = insight::get_data(model),
  criterion = mse,
  k = nrow(clusters),
  seed,
  parallel = FALSE,
  ncores = parallelly::availableCores(logical = FALSE),
  clusterVariables,
  ...
)

cv(model, data, criterion, k, seed, ...)

\method{cv}{default}(
  model,
  data = insight::get_data(model),
  criterion = mse,
  k = 10,
  seed,
  parallel = FALSE,
  ncores = parallelly::availableCores(logical = FALSE),
  ...
)

\method{print}{cv}(x, ...)

\method{cv}{lm}(
  model,
  data = insight::get_data(model),
  criterion = mse,
  k = 10,
  seed,
  method = c("auto", "hatvalues", "Woodbury", "naive"),
  parallel = FALSE,
  ncores = parallelly::availableCores(logical = FALSE),
  ...
)

\method{cv}{glm}(
  model,
  data = insight::get_data(model),
  criterion = mse,
  k = 10,
  seed,
  method = c("exact", "hatvalues", "Woodbury"),
  parallel = FALSE,
  ncores = parallelly::availableCores(logical = FALSE),
  ...
)
}
\arguments{
\item{model}{a model object that responds to model.frame(), update(), and predict()
and for which the response is stored in model$y or accessible via model.response()}

\item{data}{data frame to which the model was fit (not usually necessary)}

\item{criterion}{cross-validation criterion function of form f(y.obs, y.fitted)
(default is mse)}

\item{k}{perform k-fold cross-validation (default is 10); \code{k}
may be a number or \code{"loo"} or \code{"n"} for n-fold (leave-one-out)
cross-validation.}

\item{seed}{for R's random number generator}

\item{parallel}{do computations in parallel? (default is \code{FALSE})}

\item{ncores}{number of cores to use for parallel computations
(default is number of physical cores detected)}

\item{clusterVariables}{a character vector of names of the variables
defining clusters for a mixed model with nested random effects}

\item{...}{to match generic}

\item{x}{a \code{cv} object to be printed}

\item{method}{computational method to apply to a linear (i.e. \code{"lm"}) model
or to a generalized linear (i.e., \code{"glm"}) model. See Details for an explanation
of the available options.}
}
\value{
a "cv" object with the cv criterion averaged across the folds,
the bias-adjusted averaged cv criterion,
the criterion applied to the model fit to the full data set,
and the initial value of R's RNG seed
}
\description{
A parallelized generic k-fold (including n-fold, i.e., leave-one-out)
cross-validation function, with a default method, and
specific methods for linear and generalized-linear models that can be much
more computationally efficient.
}
\section{Methods (by class)}{
\itemize{
\item \code{cv(merMod)}: merMod method

\item \code{cv(default)}: default method

\item \code{cv(lm)}: lm method

\item \code{cv(glm)}: glm method

}}
\section{Methods (by generic)}{
\itemize{
\item \code{print(cv)}: print method

}}
\examples{
data("Auto", package="ISLR")
m.auto <- lm(mpg ~ horsepower, data=Auto)
cv(m.auto,  k="loo")
cv(m.auto, seed=1234)

data("Caravan", package="ISLR")
m.caravan <- glm(Purchase ~ ., data=Caravan[1:2500, ], family=binomial)
cv(m.caravan, k=5, criterion=BayesRule, seed=123)
}
