% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{cv}
\alias{cv}
\title{Calculate coefficient of variation (CV)}
\usage{
cv(traits)
}
\arguments{
\item{traits}{a vector of sampled trait values, minimum sample size is 10.}
}
\value{
a vector of CV estimates
}
\description{
\code{cv} estimate CV by various CV estimators from a given set of samples
}
\details{
Given a vector of sampled values, this function will return a vector of CV by six estimators.

\describe{
 \item{CV1}{the most commonly used CV estimator, \eqn{CV1=sd_sample/mu_sample},
            where sd_sampe and mu_sample are the standard deviation and mean of the sample values. }
 \item{CV2}{bias corrected CV estimator under the assumption of normal trait distribution,
             \eqn{CV2=sd_sample/mu_sample(1+1/4/N), where N is the sample size.}}
 \item{CV3}{approximate CV estimator by Breunig (2001) without assumption of sample value distribution}
 \item{CV4}{approximate CV estimator by Bao (2009) without assumption of sample value distribution}
 \item{CV5}{composite CV esetiamtor, it is the arithmetic mean of CV3 and CV4, see details in Yang (2018)}
 \item{CV6}{composite CV esetiamtor, it is the arithmetic mean of CV2 and CV4, see details in Yang (2018)}
}
}
\examples{

#simulated trait values
traits=rnorm(1000, 10, 1)

#calculate the CV
cv(traits)

}
\references{
\enumerate{
 \item Breunig, R. (2001) An almost unbiased estimator of the coefficient of variation. Economics Letters, 70 (1), 15–19.
 \item Bao, Y. (2009) Finite-sample moments of the coefficient of variation. Econometric Theory, 25 (01), 291–297.
 \item Yang et al. (2018) How to accurately estimate intraspecific trait variation? Methods in Ecology and Evolution (submitted)
}
}
