% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{cv}
\alias{cv}
\title{Calculate coefficient of variation (CV)}
\usage{
cv(traits)
}
\arguments{
\item{traits}{a vector of sampled trait values, minimum sample size is 10.}
}
\value{
a vector of CV estimates
}
\description{
Calculate coefficient of variation (CV)
}
\details{
Given a vector of sampled values, this function will return a vector of CV by six estimators.
}
\examples{

#simulated trait values
traits=rnorm(1000, 10, 1)

#calculate the CV
cv(traits)

}
\references{
Yang et al. (2018) How to accurately estimate intraspecific trait variation? Methods in Ecology and Evolution (submitted)
}
