% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCV.R
\name{FinalCompileCV}
\alias{FinalCompileCV}
\title{Compile a set of markdown documents and convert with pandoc}
\usage{
FinalCompileCV(input = c("head.md", "education.md", "employment.md",
  "publications.md", "teaching.md", "funding.md", "presentations.md"),
  output = "OMearaCV.pdf")
}
\arguments{
\item{input}{Vector of markdown documents}

\item{output}{The output file name. Pandoc will use the extension info to create a file of the right type.}
}
\description{
Compile a set of markdown documents and convert with pandoc
}

