% Generated by roxygen2 (4.0.2): do not edit by hand
\name{format_address}
\alias{format_address}
\title{Formats the address for printing}
\usage{
format_address(address)
}
\arguments{
\item{address}{a list of address lines.}
}
\value{
a character including lines breaks for TeX
}
\description{
Formats the address for printing
}

