% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCV.R
\name{GetInfoFromOrcid}
\alias{GetInfoFromOrcid}
\title{Pull in info from ORCID}
\usage{
GetInfoFromOrcid(id = "0000-0002-0337-5997")
}
\arguments{
\item{id}{Your ORCID id}
}
\value{
A list of info from ORCID
}
\description{
Pull in info from ORCID
}

