% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResponse.R
\name{getResponse}
\alias{getResponse}
\alias{getResponse.default}
\title{Extract Response Variable(s)}
\usage{
getResponse(model, ...)

\method{getResponse}{default}(model, ...)
}
\arguments{
\item{model}{a fitted model}

\item{...}{additional parameters for specific methods}
}
\value{
a vector or matrix containing the values of response variable(s)
}
\description{
Generic function to extract the response variable(s) from a fitted model.
}
\details{
The supplied \code{default} method returns the \code{model$y} component
of the model object if it exists and otherwise the result of
\code{model.response(model.frame(model))}, checking in either case whether
the result is numeric.
}
\section{Methods (by class)}{
\itemize{
\item \code{getResponse(default)}: \code{default} method

}}
\examples{
    fit <- lm(cbind(hp, mpg) ~ gear, mtcars)
    getResponse(fit)
}
