% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awards.R, R/cv-package.R, R/education.R,
%   R/employment.R, R/publications.R, R/rpackages.R, R/teaching.R
\name{list_awards}
\alias{list_awards}
\alias{list_sections}
\alias{list_education}
\alias{list_employment}
\alias{list_publications}
\alias{list_rpackages}
\alias{list_teaching}
\alias{list_curriculum_development}
\alias{list_supervision}
\title{List sections for a CV.}
\usage{
list_awards(.tbl)

list_education()

list_employment(.tbl)

list_publications(type)

list_rpackages()

list_teaching(.tbl)

list_curriculum_development(.tbl)

list_supervision(.tbl)
}
\arguments{
\item{.tbl}{CV data.}
}
\value{
Depending on output type, returns:
\itemize{
\item a \code{\link[tibble:tibble]{tibble::tibble()}} if used interactively
\item a TeX listing of the section items if used when the output document is a
PDF from the \link{vitae} package
\item a markdown listing of the section items if used when the output document
is a HTML file from the \link{pagedown} package
}
}
\description{
List sections for a CV.
}
\section{Functions}{
\itemize{
\item \code{list_awards}: List all awards given.

\item \code{list_education}: List all education taken or obtained.

\item \code{list_employment}: List all employment activities.

\item \code{list_publications}: List all publications.

\item \code{list_rpackages}: List all R packages developed (from CRAN).

\item \code{list_teaching}: List all teaching activities.

\item \code{list_curriculum_development}: List curriculum developed.

\item \code{list_supervision}: List supervisory activities.
}}

