% Generated by roxygen2 (4.0.2): do not edit by hand
\name{section_break}
\alias{section_break}
\title{Creates a section break}
\usage{
section_break(name, width = 80)
}
\arguments{
\item{name}{the name of the section}

\item{width}{the column width}
}
\value{
a character of separate lines
}
\description{
Creates a section break in LaTeX comments
}

